+++ Assets/GameAssets/Shaders/wb_shader/HitRed_dec.hlsl
@@ -1,11 +1,11 @@
 #ifndef HITRED_DEC
 #define HITRED_DEC
 
-float3 _OverlayColor;
-float _OverlayMultiple;
-float3 _HitColor; // 受击目标色
-float _HitMultiple;  // 受击目标权重
-float _HitRimPower;  // 受击边缘色参数
-float _HitRimSpread; // 受击边缘色参数
+half3 _OverlayColor;
+half _OverlayMultiple;
+half3 _HitColor; // 受击目标色
+half _HitMultiple;  // 受击目标权重
+half _HitRimPower;  // 受击边缘色参数
+half _HitRimSpread; // 受击边缘色参数
 
 #endif
